+++ internal/register/providers.go
@@ -16,6 +16,7 @@ package register
 
 import (
 	_ "github.com/coreos/ignition/v2/internal/providers/aliyun"
+	_ "github.com/coreos/ignition/v2/internal/providers/applehv"
 	_ "github.com/coreos/ignition/v2/internal/providers/aws"
 	_ "github.com/coreos/ignition/v2/internal/providers/azure"
 	_ "github.com/coreos/ignition/v2/internal/providers/azurestack"