+++ docs/release-notes.md
@@ -11,6 +11,7 @@ nav_order: 9
 
 ### Features
 
+- Support the native Apple Hypervisor
 - Support Hetzner Cloud
 - A GRUB configuration suitable for use with https://github.com/coreos/bootupd
   can now be installed; use `make install-grub-for-bootupd` to install it